Initial commit
These mtriggers have a few things that need to be cleaned up; the master
exec should be put into map order (rather than lexicographical) and some
triggers should probably be added/changed in the new system (this is a
one-to-one conversion of the old p2mtriggers). However, it should work
well enough.